requirement.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-haystack==2.8.1
geoip2==3.0.0
idna==2.10
jieba==0.42.1
maxminddb==1.5.4
PyMySQL==0.9.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
Whoosh==2.7.4
